refactor: Use new theme context
